Adding fileUpdateManager class
